+++ changelog.rst
@@ -33,6 +33,7 @@ Last change: 04-APR-2026 CEST
 - Render GUI: optimal scale bar is only set upon user's request, also in 3D
 - SPINNA allows user-defined threshold for the binary mask
 - Dialogs with scroll areas show no margins (e.g., Display settings dialog in Render)
+- Render GUI: test clustering supports G5M
 
 *Bug fixes:*
 ++++++++++++